Move and register validator
